+++ components/transactions/index.ts
@@ -1,0 +1,2 @@
+export { TransactionCard } from "./TransactionCard";
+export type { TransactionCardProps, TransactionType } from "./TransactionCard";